-fixed swipe menu (has no bugs) -refactored CRUDdb moved to kotlin -refactored fillProducts() move from DataBase to CRUDdb

+++ app/src/main/java/net/sytes/kai_soft/letsbuyka/IListActivityContract.kt
@@ -2,10 +2,11 @@ package net.sytes.kai_soft.letsbuyka
 
 import net.sytes.kai_soft.letsbuyka.Lists.List
 
+/*  Interface for activity  */
 interface IListActivityContract {
     fun onListFragmentButtonClick()
     fun onDetailFragmentButtonClick()
-    fun onListItemClick(position: Long)
+    fun onListItemClick(position: Int)
     fun onListItemLongClick(list: List)
 
 }
